backend/requirements.txt
Django
djangorestframework
django-cors-headers
drf-spectacular
requests
python-dotenv
